Add ThemesMovies component
