+++ src/containers/MoviesList.js
@@ -3,33 +3,20 @@ import PropTypes from 'prop-types';
 import spinner from './../images/spinner.gif';
 import moviesData from './moviesData';
 import Movie from './../components/Movie';
+import ThemesMovies from './../components/ThemesMovies';
 
 const MoviesList = props => {
     const [loadedChars, setLoadedChars] = useState([]);
     const [isLoading, setIsLoading] = useState(false);
-
-    //console.log(props.theme)
     
      useEffect(() => moviesData(setIsLoading, setLoadedChars, props.theme), [props.theme]);
     
       let content = <img src={spinner} alt="loading" />;
-      const themes = ['Love','Action','Sacrifice','Good','Death','Family','Reason','Faith'];
 
       if (!isLoading && loadedChars && loadedChars.length > 0) {
         content = (
           <div>
-            <span>Select a theme</span>
-            <select
-                onChange={props.onThemeSelect}
-                value={props.theme}
-                className={props.side}
-            >
-                {themes.map(theme => (
-                <option key={theme}>
-                    {theme}
-                </option>
-                ))}   
-            </select><br />
+            <ThemesMovies onThemeSelect={props.onThemeSelect} theme={props.theme} />
             <span>Production Year</span>
             <select
                 onChange={props.onCharSelect}
